+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Better bookings search allowing to search only by name
 * Removed jquery mask for birthdate
 * Better handling of birthdate (need to create another GF field, and update StormBringer theme)
+* Confirm prompt for cancellation button
 
 ### 1.4.9: June 14th, 2021
 * Removed link to dashboard from minidashboard